echo-ai-main/requirements.txt
kivy==2.2.1
pillow==10.0.0 